+++ AK/Base64.h
@@ -40,8 +40,6 @@ consteval auto base64_lookup_table()
 
 [[nodiscard]] ErrorOr<ByteBuffer> decode_base64(StringView);
 
-[[nodiscard]] ErrorOr<ByteBuffer> decode_forgiving_base64(StringView);
-
 [[nodiscard]] ErrorOr<String> encode_base64(ReadonlyBytes);
 }
