+++ Userland/Libraries/LibWeb/HTML/Window.cpp
@@ -50,6 +50,7 @@
 #include <LibWeb/HTML/WindowProxy.h>
 #include <LibWeb/HighResolutionTime/Performance.h>
 #include <LibWeb/HighResolutionTime/TimeOrigin.h>
+#include <LibWeb/Infra/Base64.h>
 #include <LibWeb/Infra/CharacterTypes.h>
 #include <LibWeb/Layout/InitialContainingBlock.h>
 #include <LibWeb/Page/Page.h>
@@ -1410,7 +1411,7 @@ JS_DEFINE_NATIVE_FUNCTION(Window::atob)
 
     // Otherwise, the user agent must convert data to a byte sequence whose nth byte is the eight-bit representation of the nth code point of data
     // and then must apply forgiving-base64 encode to that byte sequence and return the result.
-    auto decoded = AK::decode_forgiving_base64(StringView(deprecated_string));
+    auto decoded = Infra::decode_forgiving_base64(StringView(deprecated_string));
     if (decoded.is_error())
         return vm.throw_completion<JS::TypeError>(JS::ErrorType::InvalidFormat, "Base64");
